requirements_no_exr.txt
numpy==1.20.2
opencv-python==4.5.1.48
OpenEXR==1.3.2
Pillow==8.1.2
scipy==1.6.2
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.7.4.3
